# tests/snapshots/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "ClusterProfile": {
        "properties": {
          "executor_cores": {
            "title": "Executor Cores",
            "type": "integer"
          },
          "executor_instances": {
            "title": "Executor Instances",
            "type": "integer"
          },
          "executor_memory_gb": {
            "title": "Executor Memory Gb",
            "type": "integer"
          },
          "instance_hourly_rate": {
            "title": "Instance Hourly Rate",
            "type": "number"
          }
        },
        "required": [
          "executor_instances",
          "executor_cores",
          "executor_memory_gb",
          "instance_hourly_rate"
        ],
        "title": "ClusterProfile",
        "type": "object"
      },
...
      "FederationConnectRequest": {
        "properties": {
          "job_name": {
            "title": "Job Name",
            "type": "string"
          },
          "name": {
            "description": "Logical repo name for federation",
            "title": "Name",
            "type": "string"
          },
          "paths": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paths"
          },
          "ref": {
            "default": "main",
            "title": "Ref",
            "type": "string"
          },
          "subdir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subdir"
          },
          "sync": {
            "default": true,
            "title": "Sync",
            "type": "boolean"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "job_name",
          "name",
          "url"
        ],
        "title": "FederationConnectRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Data Platform Copilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/intelligence/analyze": {
      "post": {
        "operationId": "v2_intelligence_analyze",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "plan_hash",
            "required": true,
            "schema": {
              "title": "Plan Hash",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClusterProfile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/v2/intelligence/{job_name}/{intelligence_hash}": {
      "get": {
        "operationId": "v2_intelligence_get_report",
        "parameters": [
          {
            "in": "path",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "intelligence_hash",
            "required": true,
            "schema": {
              "title": "Intelligence Hash",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report",
        "tags": [
          "intelligence"
        ]
      }
    },
...
    "/federation/config": {
      "get": {
        "operationId": "get_config_federation_config_get",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Config Federation Config Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Config",
        "tags": [
          "federation"
        ]
      }
    },
    "/federation/connect": {
      "post": {
        "operationId": "connect_repo_federation_connect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FederationConnectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Connect Repo Federation Connect Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Connect Repo",
        "tags": [
          "federation"
        ]
      }
    },
    "/federation/sync": {
      "post": {
        "operationId": "sync_federation_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Req",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Sync Federation Sync Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync",
        "tags": [
          "federation"
        ]
      }
    },
...
  }
}